refactor(playground): extract usePlaygroundStoredState
